# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a POS api docs",
        "title": "POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Shabandinho",
            "email": "mahmoudshaband@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:3500",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "get All Products",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "get All Products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/output.Product"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add New Product",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Add New Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "enter product name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "enter product code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "enter product cost price",
                        "name": "cost_price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "enter product sell price",
                        "name": "sell_price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category ID",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/output.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/output.HTTPError"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "delete": {
                "description": "delete  Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "delete  Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products/{product_id}": {
            "get": {
                "description": "show  Product",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "show  Product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/output.HTTPError"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update  Product",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Update  Product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "enter product name",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "enter product code",
                        "name": "code",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "enter product cost price",
                        "name": "cost_price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "enter product sell price",
                        "name": "sell_price",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "category ID",
                        "name": "category_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Product ID",
                        "name": "product_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Product"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/output.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Category": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "category1"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Product"
                    }
                }
            }
        },
...
        "model.Inventory": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "address1"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "category1"
                },
                "productStock": {
                    "$ref": "#/definitions/model.InventoryToProduct"
                },
                "products": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Product"
                    }
                }
            }
        },
        "model.InventoryToProduct": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "buy_invoices_count": {
                    "type": "integer"
                },
                "inventory": {
                    "$ref": "#/definitions/model.Inventory"
                },
                "inventoryID": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/model.Product"
                },
                "productID": {
                    "type": "integer"
                },
                "sell_invoices_count": {
                    "type": "integer"
                }
            }
        },
        "model.Product": {
            "type": "object",
            "properties": {
                "category": {
                    "$ref": "#/definitions/model.Category"
                },
                "category_id": {
                    "type": "integer",
                    "example": 1
                },
                "code": {
                    "type": "string",
                    "example": "123"
                },
                "cost_price": {
                    "type": "string",
                    "example": "10"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "http://lorempicsm/100.jpg"
                },
                "inventories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Inventory"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "category1"
                },
                "sell_price": {
                    "type": "string",
                    "example": "100"
                }
            }
        },
...
        "output.HTTPError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "name field is required"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "output.Product": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "124"
                },
                "cost_price": {
                    "type": "string",
                    "example": "5"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image": {
                    "type": "string",
                    "example": "http://image.com/image.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "product 1"
                },
                "sell_price": {
                    "type": "string",
                    "example": "10"
                }
            }
        },
...
    }
}